move `identity_setup` to seperate file
