+++ bindings/wasm/src/account/mod.rs
@@ -6,3 +6,4 @@ mod method_secret;
 mod update;
 mod wasm_account;
 mod auto_save;
+mod wasm_identity_setup;
